requirements.txt
Django
djangohelpers
psycopg2
MySQL-python
pysqlite
git+https://github.com/django-debug-toolbar/django-debug-toolbar.git#egg=django-debug-toolbar-dev
zope.dottedname
httplib2
libopencore
topp.utils
